+++ CHANGELOG.md
@@ -117,7 +117,9 @@ Thanks to [UdHo](https://github.com/UdHo)
 
 ## 1.8.0
 
-* Added convenient type aliases for `Vector2` types: `Vec2`, `IVec2`, and `UVec2`
+* Added convenient type aliases for:
+  * `Vector2` types: `Vec2`, `IVec2`, and `UVec2`
+  * `Rectangle` types: `Rect`, `IRect`, and `URect`
 * Allow adding tuples to vectors, for example `my_vec + (1.0, 2.0)`
 
 ### New APIs